pyproject.toml
[project]
name = "safebrowsing-submit"
version = "0.1.0"
description = "Check if a URL is blacklisted by Google Safe Browsing and automatically report it using Selenium."
authors = [
    {name = "Chris",email = "goabonga@pm.me"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium (>=4.34.0,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
ruff = "^0.12.1"

[tool.poetry.scripts]
safebrowsing-submit = "safebrowsing_submit.main:main"